feat: Add template download to relational import jobs Fixes: SIRI-1194
