+++ src/main/java/sirius/biz/jobs/batch/file/EntityImportJobFactory.java
@@ -8,24 +8,17 @@
 
 package sirius.biz.jobs.batch.file;
 
-import org.apache.commons.io.output.ByteArrayOutputStream;
 import sirius.biz.importer.ImportContext;
 import sirius.biz.importer.Importer;
-import sirius.biz.importer.format.FieldDefinition;
 import sirius.biz.importer.format.ImportDictionary;
 import sirius.biz.jobs.infos.JobInfoCollector;
 import sirius.biz.jobs.params.Parameter;
 import sirius.biz.process.ProcessContext;
 import sirius.db.mixing.BaseEntity;
-import sirius.kernel.commons.CSVWriter;
 import sirius.kernel.commons.Explain;
 import sirius.kernel.health.Exceptions;
 import sirius.kernel.health.Log;
 
-import javax.annotation.Nullable;
-import java.io.OutputStreamWriter;
-import java.nio.charset.StandardCharsets;
-import java.util.Base64;
 import java.util.function.Consumer;
 
 /**
@@ -112,26 +105,6 @@ public abstract class EntityImportJobFactory extends DictionaryBasedImportJobFac
         super.collectJobInfos(collector);
         collector.addTranslatedCard("EntityImportJobFactory.automaticMappings");
         getDictionary().emitJobInfos(collector);
-    }
-
-    @Override
-    @Nullable
-    public String generateTemplateUrl() {
-        try (ByteArrayOutputStream out = new ByteArrayOutputStream();
-             OutputStreamWriter writer = new OutputStreamWriter(out, StandardCharsets.UTF_8);
-             CSVWriter csvWriter = new CSVWriter(writer)) {
-            csvWriter.writeUnicodeBOM();
-            Object[] headers = getDictionary().getFields().stream()
-                                              .filter(field -> !field.isHidden())
-                                              .map(FieldDefinition::getLabel)
-                                              .toArray();
-            csvWriter.writeArray(headers);
-            writer.flush();
-            return "data:text/csv;base64," + Base64.getEncoder().encodeToString(out.toByteArray());
-        } catch (Exception exception) {
-            Exceptions.handle(Log.BACKGROUND, exception);
-            return null;
-        }
     }
 
     @Override
